# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能",
        "title": "MiMusic API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mimusic-org/mimusic",
            "email": "im.hanxi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.4.1"
    },
    "host": "localhost:58091",
    "basePath": "/api/v1",
    "paths": {
...
        "/upgrade/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "检查是否有可用的新版本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统升级"
                ],
                "summary": "检查更新",
                "responses": {
                    "200": {
                        "description": "成功返回更新检查结果",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "非 Docker 环境不支持升级",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "检查更新失败",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upgrade/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前升级任务的进度信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统升级"
                ],
                "summary": "获取升级进度",
                "responses": {
                    "200": {
                        "description": "成功返回升级进度",
                        "schema": {
                            "$ref": "#/definitions/models.UpgradeProgress"
                        }
                    },
                    "403": {
                        "description": "非 Docker 环境不支持升级",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upgrade/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将二进制文件回退到 Docker 镜像中的原始版本，然后重启服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统升级"
                ],
                "summary": "回退到底包版本",
                "responses": {
                    "200": {
                        "description": "回退已开始",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "403": {
                        "description": "非 Docker 环境不支持回退",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upgrade/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "开始升级到指定版本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统升级"
                ],
                "summary": "开始升级",
                "parameters": [
                    {
                        "description": "升级请求 {version_type: stable|dev}",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "升级已开始",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "非 Docker 环境不支持升级",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "升级失败",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upgrade/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取正式版和测试版的版本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统升级"
                ],
                "summary": "获取可用版本信息",
                "responses": {
                    "200": {
                        "description": "成功返回版本信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "非 Docker 环境不支持升级",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "获取版本信息失败",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "详细错误信息（可选）",
                    "type": "string",
                    "example": "详细错误信息"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "操作失败"
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "成功信息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
        "models.UpgradeProgress": {
            "type": "object",
            "properties": {
                "current_step": {
                    "description": "当前步骤描述",
                    "type": "string",
                    "example": "正在下载新版本..."
                },
                "error": {
                    "description": "错误信息（如有）",
                    "type": "string",
                    "example": "下载失败"
                },
                "progress": {
                    "description": "进度百分比 (0-100)",
                    "type": "integer",
                    "example": 50
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "enum": [
                        "idle",
                        "downloading",
                        "testing",
                        "replacing",
                        "restarting",
                        "completed",
                        "failed"
                    ],
                    "example": "downloading"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}